requirements.txt
certifi==2018.8.13
chardet==3.0.4
cycler==0.10.0
defusedxml==0.5.0
Django==1.11
django-allauth==0.36.0
django-braces==1.13.0
django-crispy-forms==1.7.2
django-model-utils==3.1.2
django-mysql==2.4.1
django-paypal==0.5.0
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-rest-passwordreset==0.9.5
django-simple-poll==0.2.3
django-star-ratings==0.7.0
django-tables2==2.0.1
django-vote==2.1.7
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
idna==2.7
image==1.5.24
kiwisolver==1.0.1
matplotlib==2.2.3
numpy==1.15.1
oauthlib==2.1.0
pdf2image==0.1.14
Pillow==5.2.0
poll==1.0
PyJWT==1.6.4
pylovepdf==1.3.1
pyparsing==2.2.1
PyPDF2==1.26.0
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.5
reportlab==3.5.6
requests==2.19.1
requests-oauthlib==1.0.0
rest-social-auth==1.4.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
swapper==1.1.0
urllib3==1.23
